move WrappingBehavior to internal
