+++ akka-actor-typed/src/main/scala/akka/actor/typed/Behavior.scala
@@ -9,6 +9,7 @@ import akka.actor.typed.internal.BehaviorImpl
 import scala.annotation.tailrec
 
 import akka.actor.typed.internal.BehaviorImpl.OrElseBehavior
+import akka.actor.typed.internal.WrappingBehavior
 import akka.util.{ LineNumbers, OptionVal }
 import akka.annotation.{ DoNotInherit, InternalApi }
 import akka.actor.typed.scaladsl.{ ActorContext ⇒ SAC }
